Move align module to its own package
